+++ spec/lib/html2rss/selectors/categories_extractor_spec.rb
@@ -3,10 +3,6 @@
 require 'spec_helper'
 
 RSpec.describe Html2rss::Selectors::CategoriesExtractor do
-  subject(:extractor) do
-    described_class.new(config_lookup:, attribute_selector: Html2rss::Selectors::AttributeSelector.new)
-  end
-
   let(:selectors_config) do
     {
       items: { selector: 'article', enhance: false },
@@ -14,15 +10,6 @@ RSpec.describe Html2rss::Selectors::CategoriesExtractor do
       tags: { selector: '.tags a', extractor: 'text' },
       categories: %i[category tags]
     }
-  end
-  let(:config_lookup) do
-    lambda do |name, allow_nil: false|
-      key = name.to_sym
-      return [key, selectors_config[key]] if selectors_config.key?(key)
-      return [key, nil] if allow_nil
-
-      raise Html2rss::Selectors::InvalidSelectorName, "Selector for '#{key}' is not defined."
-    end
   end
   let(:response) do
     Html2rss::RequestService::Response.new(
@@ -48,14 +35,14 @@ RSpec.describe Html2rss::Selectors::CategoriesExtractor do
     )
   end
 
-  describe '#select_categories' do
+  describe '.select_categories' do
     it 'flattens single- and multi-node category selectors into discrete strings' do
-      expect(extractor.select_categories(category_selectors: %i[category tags], scope:))
+      expect(described_class.select_categories(category_selectors: %i[category tags], scope:))
         .to eq(%w[News Ruby RSS])
     end
 
     it 'returns an empty list when a referenced selector is missing' do
-      expect(extractor.select_categories(category_selectors: %i[missing], scope:)).to eq([])
+      expect(described_class.select_categories(category_selectors: %i[missing], scope:)).to eq([])
     end
   end
 end